Update strategies modules to be more organised and more modular for future extensions
