--- bbstrader/strategies/.docs/Arch.md
@@ -1,140 +1,0 @@
-
-# ArimaGarchStrategy Class Documentation
-
-The `ArimaGarchStrategy` class implements a trading strategy by combining ARIMA (AutoRegressive Integrated Moving Average) and GARCH (Generalized Autoregressive Conditional Heteroskedasticity) models. This strategy aims to predict future returns based on historical price data and is built upon a generic Strategy class, which it extends with specific strategy logic.
-
-The strategy is implemented in the following steps:
-1. Data Preparation: Load and prepare the historical price data.
-2. Modeling: Fit the ARIMA model to the data and then fit the GARCH model to the residuals.
-3. Prediction: Predict the next return using the ARIMA model and the next volatility using the GARCH model.
-4. Trading Strategy: Execute the trading strategy based on the predictions.
-5. Vectorized Backtesting: Backtest the trading strategy using the historical data.
-
-## Class Attributes
-
-- `symbol` (str): The ticker symbol for the financial instrument being analyzed.
-- `data` (pd.DataFrame): The dataset prepared for model training and prediction, including log returns and possibly differenced log returns.
-- `k` (int): The window size for rolling prediction, indicating how many past observations are considered for each prediction.
-
-## Methods
-
-### `__init__(self, symbol, data, k)`
-
-Initializes the `ArimaGarchStrategy` class with a symbol, raw data, and window size.
-
-#### Parameters
-
-- `symbol` (str): The ticker symbol for the financial instrument.
-- `data` (pd.DataFrame): The raw dataset containing at least 'Close' prices.
-- `k` (int): The window size for rolling prediction.
-
-### `load_and_prepare_data(self, df)`
-
-Prepares the dataset by calculating logarithmic returns and differencing if necessary.
-
-#### Parameters
-
-- `df` (pd.DataFrame): The raw dataset containing at least 'Close' prices.
-
-#### Returns
-
-- `pd.DataFrame`: The dataset with additional columns for log returns and differenced log returns.
-
-### `fit_best_arima(self, window_data)`
-
-Fits the ARIMA model to the provided window of data, selecting the best model based on AIC.
-
-#### Parameters
-
-- `window_data` (pd.DataFrame): The dataset for a specific window period.
-
-#### Returns
-
-- ARIMA model: The best fitted ARIMA model based on AIC.
-
-### `show_arima_garch_results(self, arima_result, acf=True, test_resid=True)`
-
-Displays the ARIMA and GARCH model results, including plotting ACF of residuals and conducting Box-Pierce and Ljung-Box tests.
-
-#### Parameters
-
-- `arima_result` (ARIMA model): The ARIMA model result.
-- `acf` (bool, optional): If True, plot the ACF of residuals. Defaults to True.
-- `test_resid` (bool, optional): If True, conduct Box-Pierce and Ljung-Box tests on residuals. Defaults to True.
-
-### `fit_garch(self, window_data)`
-
-Fits the GARCH model to the residuals of the best ARIMA model.
-
-#### Parameters
-
-- `window_data` (pd.DataFrame): The dataset for a specific window period.
-
-#### Returns
-
-- tuple: Contains the ARIMA result and GARCH result.
-
-### `predict_next_return(self, arima_result, garch_result)`
-
-Predicts the next return using the ARIMA model and the next volatility using the GARCH model.
-
-#### Parameters
-
-- `arima_result` (ARIMA model): The ARIMA model result.
-- `garch_result` (GARCH model): The GARCH model result.
-
-#### Returns
-
-- float: The predicted next return.
-
-### `get_prediction(self, window_data)`
-
-Generates a prediction for the next return based on a window of data.
-
-#### Parameters
-
-- `window_data` (pd.DataFrame): The dataset for a specific window period.
-
-#### Returns
-
-- float: The predicted next return.
-
-### `calculate_signals(self, window_data)`
-
-Calculates the trading signal based on the prediction: 'LONG' if positive, 'SHORT' if negative.
-
-#### Parameters
-
-- `window_data` (pd.DataFrame): The dataset for a specific window period.
-
-#### Returns
-
-- str: The trading signal ('LONG', 'SHORT', or None).
-
-### `execute_trading_strategy(self, predictions)`
-
-Executes the trading strategy based on a list of predictions, determining positions to take.
-
-#### Parameters
-
-- `predictions` (list): A list of predicted returns.
-
-#### Returns
-
-- list: A list of positions (1 for 'LONG', -1 for 'SHORT', 0 for 'HOLD').
-
-### `backtest_strategy(self)`
-
-Performs a backtest of the strategy over the entire dataset, plotting cumulative returns.
-
-### `plot_cumulative_returns(self, strategy_returns, buy_and_hold_returns, dates)`
-
-Plots the cumulative returns of the ARIMA+GARCH strategy against a buy-and-hold strategy.
-
-#### Parameters
-
-- `strategy_returns` (np.array): Cumulative returns from the strategy.
-- `buy_and_hold_returns` (np.array): Cumulative returns from a buy-and-hold strategy.
-- `dates` (pd.Index): The dates corresponding to
-
- the returns.
