--- bbstrader/strategies/.docs/Klfiter.md
@@ -1,107 +1,0 @@
-# KLFStrategy Class Documentation
-
-The `KLFStrategy` class implements a trading strategy based on the Kalman Filter. This filter is a recursive algorithm used for estimating the state of a linear dynamic system from a series of noisy measurements. The strategy is designed to process market data, estimate dynamic parameters such as the slope and intercept of price relationships, and generate trading signals based on those estimates.
-
-## Initialization
-
-```python
-def __init__(self, tickers: list | tuple, **kwargs):
-```
-
-Initializes the Kalman Filter strategy with the given tickers and optional parameters.
-
-### Parameters
-
-- `tickers`: A list or tuple of ticker symbols representing financial instruments.
-- `**kwargs`: Keyword arguments for additional parameters. Accepts `delta` and `vt` as optional parameters.
-
-## Public Methods
-
-### `_init_kalman`
-
-```python
-def _init_kalman(self):
-```
-
-Initializes and returns a Kalman Filter configured for the trading strategy. The filter is set up with initial state and covariance, state transition matrix, process noise, and measurement noise covariances.
-
-### `calc_slope_intercep`
-
-```python
-def calc_slope_intercep(self, prices: np.ndarray):
-```
-
-Calculates and returns the slope and intercept of the relationship between the provided prices using the Kalman Filter. This method updates the filter with the latest price and returns the estimated slope and intercept.
-
-#### Parameters
-
-- `prices`: A numpy array of prices for two financial instruments.
-
-#### Returns
-
-- A tuple containing the slope and intercept of the relationship.
-
-### `calculate_xy_signals`
-
-```python
-def calculate_xy_signals(self, et, std):
-```
-
-Generates trading signals based on the forecast error and standard deviation of the predictions. It returns signals for exiting, going long, or shorting positions based on the comparison of the forecast error with the standard deviation.
-
-#### Parameters
-
-- `et`: The forecast error.
-- `std`: The standard deviation of the predictions.
-
-#### Returns
-
-- A tuple containing the trading signals for the two financial instruments.
-
-### `calculate_signals`
-
-```python
-def calculate_signals(self, prices: np.ndarray):
-```
-
-Calculates trading signals based on the latest prices and the Kalman Filter's estimates. It updates the filter's state with the latest prices, computes the slope and intercept, and generates trading signals based on the forecast error and prediction standard deviation.
-
-#### Parameters
-
-- `prices`: A numpy array of prices for two financial instruments.
-
-#### Returns
-
-- A dictionary containing trading signals for the two financial instruments.
-
-## Attributes
-
-- `tickers`: The list or tuple of ticker symbols for the financial instruments.
-- `latest_prices`: A numpy array storing the latest prices for the financial instruments.
-- `delta`: A small constant used in the calculation of process noise covariance.
-- `wt`: Process noise covariance matrix.
-- `vt`: Measurement noise variance.
-- `theta`: A numpy array storing the estimated slope and intercept.
-- `P`: The error covariance matrix.
-- `R`: Measurement noise covariance.
-- `kf`: The Kalman Filter instance used in the strategy.
-
-This class provides a comprehensive framework for implementing a trading strategy based on the Kalman Filter, allowing for the dynamic estimation of market parameters and the generation of trading signals based on those estimates.
-
-## Example
-
-```python
-# Define the tickers
-tickers = ('AAPL', 'MSFT')
-
-# Initialize the KLFStrategy with tickers and optional parameters
-klf_strategy = KLFStrategy(tickers, delta=1e-4, vt=1e-3)
-
-# Example prices for AAPL and MSFT
-prices = np.array([150.0, 250.0])
-
-# Calculate trading signals based on current prices
-signals = klf_strategy.calculate_signals(prices)
-
-print(signals)
-````
